requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
Cython==3.0.10
dash==2.17.1
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-mantine-components==0.10.2
dash-table==5.0.0
decorator==5.1.1
distro==1.9.0
ecos==2.0.13
executing==2.0.1
Flask==3.0.3
Flask-Login==0.6.3
fonttools==4.53.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==7.1.0
ipython==8.25.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mlxtend==0.23.1
nest-asyncio==1.6.0
numexpr==2.10.0
numpy==1.26.4
openai==1.34.0
osqp==0.6.7
packaging==24.1
pandas==2.2.2
parso==0.8.4
patsy==0.5.6
pillow==10.3.0
plotly==5.22.0
pmdarima==2.0.4
prompt_toolkit==3.0.47
pure-eval==0.2.2
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
qdldl==0.1.7.post2
requests==2.32.3
retrying==1.3.4
scikit-learn==1.3.2
scikit-survival==0.22.2
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
statsmodels==0.14.2
tenacity==8.3.0
threadpoolctl==3.5.0
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
waitress==3.0.0
wcwidth==0.2.13
Werkzeug==3.0.3
zipp==3.19.2
